webpack.config.js: use same rules for css and scss
This fixes a bug in production builds where CSS files (such as
normalize.css) and our own CSS-compiled SCSS files are loaded out of
order. By using the same rule for both CSS and SCSS files, all CSS files
are fed to mini-css-extract-plugin and bundled in the order they are
used in index.js.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>