Defer fade-in animation after initial page load
Sapper's export functionality crawls the loaded page to figure out which
static pages to generate. If we initially hide the content of the page
in _layout.svelte, the page appears empty. This means that blog posts
are not exported cause there are no links initially pointing to them
when the front page loads. To circumvent the issue, we only fade-in
pages after the initial page load.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>